Move s3 logic to new module routes/s3api.js
